v1: add path semantics check stub to BaseProvider
Relevant: [OSF-4675]

v1 expects folder paths to always end with a slash and file paths to
never end with a slash.  However, it wasn't enforcing this and some
manually-constructed tests were 500ing as a result.  It was decided
that trying to address folders as file and vice-versa should
consistently produce 404s.  This is especially important for providers
like GitHub, where a file named foo and a directory named foo could
feasibly coexist.