github: fix move to always remove src file
When moving files from a non-root folder in github, the source file was
being left behind.  This was because the tree entry for the source
directory was left in the modified master tree, where it still contained
references to the old blob/tree.  The solution was to remove all tree
entries from the master tree and to allow github to infer the new tree
structure from the path names of the remaing blobs.

Fixes: [#OSF-4792]