github: avoid dataloss w/ folderops in large repos
Folders (trees) in GitHub cannot be operated on directly, and must
instead be moved/renamed/deleted by manipulating the blob paths in the
master tree.  Unfortunately, GH truncates master tree listings for repos
with lots of files and does not provide a way to page through and
retrieve all entries. Any blobs omitted when updating the master tree
will be treated as deleted.  Therefore, it doesn't seem that we can
always safely perform certain folder ops on large repos, so instead we
throw a 501 Not Implemented with an explanatory message.

Fixes [#OSF-4867]