github: fix created semantics for folder move/copy
The logic for determining whether a folder that was moved or copied was
incorrect. Github tree paths do not have trailing slashes, so exists was
always being set to false.  In addition, the caller expects the second
returned arg from this method to be created, which is the negation of
exists.

Fixes: [#OSF-4870]