s3: implement recursive delete
A DELETE request against a non-empty folder path on Amazon S3 will not
delete the folder, because folders on S3 are inferred from the pathnames
of the file objects.  To fully delete a directory, we must delete each
object within it.  Amazon provides a bulk delete operation to make this
more efficient.

To run the bulk delete, we must first get the keys of all the objects
within the directory, then POST a request with an XML body containing
the keys to the delete endpoint.  S3 limits their object list
responses and bulk delete requests to 1000 entries.  In the event that
more than 1000 entries need to be deleted, we issue multiple requests
in 1000-key batches.

Fixes: [#OSF-4881]