osfstorage: break out signed url build into a method
osfstorage signs its requests by adding a query parameter with the
signature of the payload.  That signature calculating code was built
into the request issuing code, making it hard to test.  This commit
splits out that code into a new method, `build_signed_url`, that
returns the request with the parameters added.  `make_signed_request`
calls this function. A previously failing test calls
`build_signed_url` to calculate the correct url to intercept for
response-mocking purposes.