v1: make self.path always refer to url entity
For most actions self.path will always refer to the entity described by
the URL, but for the create subfolder and create new file actions,
self.path is modified to point to the new entity.  This could cause
confusion, so create a new target_path attribute to hold the new entity
path.  For the file update actions, self.path == self.target_path.