fix modified timestamps from filesystem provider
When running a local dev copy of WB, the modified timestamp returned
from file metadata was for the local time, and lacked an explicit
timezone attachment.  When the file metadata is stored in mongo, mongo
assumes it is Zulu time.  This was causing erroneous and mismatched
modified and created times for osfstorage.

This branch updates filesystem to return a utc datetime with an active
timezone object.  It also updates two tests that weren't being explicit
about their timezone.  Those tests worked fine without the tzinfo, but
we need to start being more disciplined about this.