update s3 provider to work w/ sigv4 regions
The Frankfort and Seoul AWS regions require that requests be signed with
their v4 protocol.  The code to handle this is in boto, but the
generate_url_sigv4 function doesn't include query parameters in the
signing protocol.  That means that when we issue a request with query
parameters, the signature we report fails to match with that calculated
by Amazon.

I've updated the boto code to handle this and updated the s3 provider to
always pass all query parameters and headers to the generate_url
functions.  sigv2 allowed some to be omitted, but v4 is much stricter.