fix tornado setting status code to 200 for 404s
There is a bug in tornado where setting the status using the http.client
status code was causing it to always return 200 instead of the actual
status code.  There is a patch for this in the unreleased
tornado (b507c9ac) from our very own @chrisseto.  Until a new tornado is
released, work around this by explicitly casting the status code as an
int().  This bug was only affecting v1 queries.