handle exists == [] in renaming logic
Annoyingly, the `exists()` method returns the result of `metadata()` if
the queried entity exists.  For an empty directory, this response will
be `[]` which tests as false.  We need to make sure when we call
`exists()` that we explicitly check whether the return value is `[]` and
if so, treat it as True.