move v1 PUT validation into postvalidate_put()
When I added strict PUT validation to V1, I put the bulk of the code in
the `.put()` method.  But by the time the code has reached that point,
the upload has already kicked off.  The fix is to move the validation
back into `postvalidate_put()` so we can abort before starting the
upload.