github: teach move/copy folders to overwrite
When moving and copying folders in github, the contents of the source
were being merged with the contents of the destination.  Move/copy
should overwrite; if the user has requested to keep both, the
`dest_path` received by `_intra_copy_or_move` will reflect the new name.

To do this, make sure we prune the destination path entries from the
tree if the destination path is a folder.