github: enable v1 ops for files larger than 1Mb
The github provider's `validate_v1_path` method was using the Github
`contents` API endpoint to retrieve metadata about the path, so that
it could determine whether a file or folder at that path already
exists and is of the correct type.  Unfortunately, the `contents`
endpoint also returns the file contents along with the metadata.
Because of this, GitHub will refuse to respond if the requested file
is bigger than 1Mb.  `validate_v1_path` has been updated to request a
recursive tree listing, which it can then search through for the given
path.

Fixes: [#OSF-5866]