github: throw 400 if multiple branches are given
The query argument processing code will pass the branch kwarg as a list
if the same param appears multiple times in a query string.  Passing a
list as the ref to operate on makes no sense, so the github provider
will now throw a 400 Bad Request error if multiple args are given.
e.g.:

  GET /resource/mst3k/provider/github/foo?meta=&branch=master&branch=develop

h/t @TomBaxter