centralize callback logging
WB logs mutating actions to a callback given by the auth provider.
This logging was happening in five different places in the code:

 * .server.api.v0.core.BaseProviderHandler._send_hook
 * .server.api.v0.core.BaseCrossProviderHandler._send_hook
 * .server.api.v1.provider.__init__.ProviderHandler._send_hook
 * .task.copy.copy
 * .task.move.move

This commit moves all logging code to one subroutine,
.core.utils.log_to_callback.  This subroutine takes a source LogPayload
and an optional destination LogPayload and constructs a common log
payload to be sent to the logging server.

The LogPayload class is a simple wrapper class to hold the logged
object's parent resource, serving provider, and a path or metadata
object to be serialized.  It serves to enforce the argument contract and
to format the log object in a common way.