Removed 3 functions that are no longer in use
Removed _materialized_path_to_id. It was added and seemed to never be used, or if it was, it was removed and rewritten long ago

Removed _resolve_id_to_parts. This was once used in validate_path, but was commented out and removed at one point. no longer in use

Removed _get_parent_ids. was only used by the above function.