Handle GitHub rate limit during inter provider move/copy
[SVCS-543]

  The default GitHub rate limit is 5,000 calls per hour. The hour does
not float. 5,000 calls is insufficient to accommodate the registration of
repos with more then ~1,200 files. Therefore this PR self rate limits at
the GitHub provider's make_request method.
  This PR only rate limits GitHub make_requests that run in celery
tasks. Currently, the only tasks are "move" and "copy". The distinction
is made by adding a "task_id" attribute to the providers after they are
instantiated in the task definition.
  This PR uses the rate limit information returned in all GitHub
response headers to limit the number of make_requests in such a way that
the GitHub rate limit is never reached. In practice this usually results
in a rate limit between 1-2 calls a second.
  As a bonus, this PR adds a GitHubRateLimitExceeded exception, that
will be checked for with each make_request. This exception should never
be raised during a move/copy celery task, but may be useful for
non-celery make_requests.
  In addition to adding tests for rate limiting, this PR cleans up two
unnecessary aysnc defs that were causing "was never awaited"
RuntimeWarnings during pytest runs.