Use Boto3
Upgrades boto2 to boto3.
This required some implementations of operations to change because of
some inflexibility inside of boto - It likes to do it the boto way, but
using some of the internals, to generate signed urls for example, is
difficult to call in isolation. In the interest of getting it
implemented sooner than later, we use the boto way of doing some things
rather than making our own requests manually.