Upgrade tornado
The latest version of tornado has updates to use the event loop properly
that we need for aiohttp context managers to work properly. The version
of tornado that we use doesn't use the asyncio event loop, but its own
event loop, so other libraries that use the event loop sometimes don't
work properly. Like aiohttp.