Manually update the reset time when WB detects the time has come
Normally, the reset time is only updated with the response headers.
However, when reaching the limit, requests are made at a super slow
rate (1 per 100 seconds with default settings) so the update has an
100 seconds delay. Now WB just adds one hour to the reset time when
it detects that the current time has passed the current reset time
so new requests can be allowed as soon as possible.