Fixed revisions for Bitbucket API 2.0
Use the file history endpoint to obtain the commit list. Then use
the commit url for each to get the commit metadata. From WB's
perspective, a few things are no longer available in BB API 2.0.
e.g the branch of the commit, the history before a revived file
was deleted, etc.