Fixed an issue where branch name containing a slash breaks BB API
Endpoint /2.0/repositories/{username}/{repo_slug}/src/{node}/{path}
returns 404 if the node segment is a branch name containing a slash.
This is a limitation/bug on several BB API 2.0 endpoints, which has
nothing to do with encoding. Below is the closest reference we found:
https://bitbucket.org/site/master/issues/9969/get-commit-revision-api-does-not-accept

The fix is simple: make an extra request to fetch the commit sha of
the branch and update the path object. Must apply this before using
path.ref to building a BB API URL (3 occurrences in the BB provider).