Add obfuscation to redis keys for rate-limiting
OSF cookie, OAuth access token and base64-encoded username/password
are used as redis keys during rate-limiting. We must obfuscate them
for the same reason we store only password hashes in database. The
`hashlib.hash256().hexdigest()` is used in this case. In addition,
a prefix is added to the digest to identifty which type it is. The
"No Auth" case is hashed as well (though unnecessarily) so that the
keys all have the same look and length.