Fix request auth for Dataverse SWORD API using aiohttp3
aiohttp 3.5 (more specifically, starting aiohttp 0.22) requires
the req auth param to be of type BasicAuth when building the req
while aiohttp 0.18 politely shows a deprecation warning when a
tuple is used. The fix is to simply use a BasicAuth object init
with the tuple that was used as the req auth param before.

Only upload and delete are updated since they are the only two
actions that use the SWORD API.