WB client session now supports customized connector
For providers that use a customized connector such as owncloud,
the new session is created (if no session is found in the event
loop) with the given connector while an existing session simply
ignores (and closes) the new connector. Given that the session
is per loop for each instance, the existing session if found
must already have a connector with qualified customizations.