Fix missing headers when registering multiple URLs in GitHub tests
aiohttp3 is more strict on response content-type, where .json()
would fail if the content-type were not provided or were incorrect.
However, the updated aiohttpretty still has the unpleasant bug that
.register_uri() ignores headers provided by .register_json_uri()
when multiple URL registrations are used. The fix is to explicitly
add the required header when calling .register_json_uri().