Fix download and tests for GitLab, update docstr
The GL provider has been using a complicated hack for downloads due to
this issue: https://gitlab.com/gitlab-org/gitlab-ce/issues/31470,
which breaks the dedicated endpoint for direct raw file download.
Fortunately, the bug was fixed in GLv10.0 so that download has been
written to use the preferred and intuitive approach.

In addition, the hack would not work with aiohttp3 even if we kept it.
It needs to modify response headers directly within the object before
passing it to the response stream reader. aiohttp3 makes this
impossible: the `headers` attr is defined as a `@reify` property.  The
cache effect of `@reify` prevents modifying either `_headers` or
`headers`.